+++ lib/hexapdf/layout.rb
@@ -44,6 +44,7 @@ module HexaPDF
     autoload(:InlineBox, 'hexapdf/layout/inline_box')
     autoload(:LineFragment, 'hexapdf/layout/line_fragment')
     autoload(:TextShaper, 'hexapdf/layout/text_shaper')
+    autoload(:TextBox, 'hexapdf/layout/text_box')
 
   end